% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Hybrid_Points.R
\name{Extract_HybridV2_Points}
\alias{Extract_HybridV2_Points}
\title{Extract_Hybrid_V2_Points}
\usage{
Extract_HybridV2_Points(
  List.Points,
  Project.Name,
  Path.Hybrid,
  Target.Met.Var = c("pr", "tasmin", "tasmax", "sfcWind", "hurs", "rsds", "rlds"),
  Start.Year = 1950,
  End.Year = 2019,
  Output.Path,
  File.Format = 1
)
}
\arguments{
\item{List.Points}{A array or matrix of points:  Lon and Lat in degrees}

\item{Project.Name}{Basin (or Project) name}

\item{Path.Hybrid}{String of a folder path that hybrid V2 data (RData format) are available}

\item{Target.Met.Var}{Meteorological variables to be extracted, data(Met.Var.Name). Ex: c('pr', 'tmin', 'tmax', 'windspeed', 'relhumid')}

\item{Start.Year}{Start year to extract the hybrid climate dataset (default=1950)}

\item{End.Year}{Ending year to extract the hybrid climate dataset(default=2019)}

\item{Output.Path}{Output path to store outputs (e.g., Output)}

\item{File.Format}{File format of extracted data; 1:RDS (default) 2:RData 3: csv}
}
\value{
Location data (Lat, Lon and Elevation), Meteorological variables within target domain in rds format
}
\description{
Extracting Hybrid V2 at specific points (extracting from the nearest grid cells)
}
\references{
Eum, H.-I., Gupta, A., 2019. Hybrid climate datasets from a climate data evaluation system and their impacts on hydrologic simulations for the Athabasca River basin in Canada. Hydrology and Earth System Sciences 23, 5151–5173. https://doi.org/10.5194/hess-23-5151-2019

Eum, H.-I., Fajard, B., Tang, T., Gupta, A., 2023. Potential changes in climate indices in Alberta under projected global warming of 1.5–5 °C. Journal of Hydrology: Regional Studies 47, 101390. https://doi.org/10.1016/j.ejrh.2023.101390
}
