% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Hybrid_SHP.R
\name{Extract_HybridV2_SHP}
\alias{Extract_HybridV2_SHP}
\title{Extract_SHP}
\usage{
Extract_HybridV2_SHP(
  Path.SHP,
  Basin.Name,
  Path.Hybrid,
  Target.Met.Var,
  Buffer = 5000,
  Start.Year = 1950,
  End.Year = 2019,
  Output.Path,
  File.Format = 1
)
}
\arguments{
\item{Path.SHP}{Path and file name of a shape file}

\item{Basin.Name}{Basin (or Project) name}

\item{Path.Hybrid}{String of a folder path that hybrid V2 data (RData format) are available}

\item{Target.Met.Var}{Meteorological variables to be extracted, data(Met.Var.Name). Ex: c('pr', 'tmin', 'tmax', 'windspeed', 'relhumid')}

\item{Buffer}{Buffer in meter (default=5000)}

\item{Start.Year}{Start year to extract the hybrid climate dataset (default=1950)}

\item{End.Year}{Ending year to extract the hybrid climate dataset(default=2019)}

\item{Output.Path}{Output path to store outputs}

\item{File.Format}{File format of extracted data; 1:RDS (default) 2:RData 3: csv}
}
\value{
Location data (Lat, Lon and Elevation), Precipitation, (Min/Max) Temperature
}
\description{
Extracting the hybrid V2 climate dataset within a boundary of a shape file
The boundary information is obtained from a shape file
}
\examples{
Extract_Hybrid_SHP("Input/SSRB.shp","SSRB","Input/PRCP.RData","Input/Tmin.RData","Input/Tmax.RData",
Buffer=1.0,Start.Year=1950,End.Year=2019,Output.Path="Output")
}
\references{
Eum, H.-I., Gupta, A., 2019. Hybrid climate datasets from a climate data evaluation system and their impacts on hydrologic simulations for the Athabasca River basin in Canada. Hydrology and Earth System Sciences 23, 5151–5173. https://doi.org/10.5194/hess-23-5151-2019
}
